import FoundationNetwork
collected all the work with FoundationNetwork in one file
